chore: Move metastore related protos and types to separate crates
Doing this to resolve a dependency cycle for the table funcs that are being
worked on.

The idea is that crates that depend on types from metastore shouldn't depend on
the service impl too. This was already true in our case, so the refactor was
pretty straightforward.

`datasources` still depends on `metastore`, but I'm going to open up a followup
pr that adds a `metastorebuiltins` crate which will contain the stuff that
`datasources` currently depends on.
