+++ crates/datasources/src/debug/mod.rs
@@ -23,7 +23,7 @@ use datafusion::physical_plan::{
 };
 use errors::DebugError;
 use futures::Stream;
-use metastore::types::options::TunnelOptions;
+use metastoreproto::types::options::TunnelOptions;
 use serde::{Deserialize, Serialize};
 use std::any::Any;
 use std::fmt;
