+++ crates/datasources/src/mysql/mod.rs
@@ -27,7 +27,7 @@ use datafusion::physical_plan::{
     ExecutionPlan, Partitioning, RecordBatchStream, SendableRecordBatchStream, Statistics,
 };
 use futures::{Stream, StreamExt, TryStreamExt};
-use metastore::types::options::TunnelOptions;
+use metastoreproto::types::options::TunnelOptions;
 use mysql_async::consts::{ColumnFlags, ColumnType};
 use mysql_async::prelude::Queryable;
 use mysql_async::{
