+++ crates/glaredb/Cargo.toml
@@ -16,6 +16,7 @@ pgsrv = {path = "../pgsrv"}
 pgrepr = {path = "../pgrepr"}
 object_store = {version = "0.5", features = ["gcp"]}
 metastore = {path = "../metastore"}
+metastoreproto = { path = "../metastoreproto" }
 anyhow = "1.0"
 tokio = { version = "1", features = ["full"] }
 clap = { version = "4.2.7", features = ["derive"] }
