+++ crates/glaredb/src/util.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
 use metastore::local::{start_inprocess_inmemory, start_inprocess_local};
-use metastore::proto::service::metastore_service_client::MetastoreServiceClient;
+use metastoreproto::proto::service::metastore_service_client::MetastoreServiceClient;
 use std::fs;
 use std::path::{Path, PathBuf};
 use tonic::transport::Channel;
