+++ crates/metastore/src/lib.rs
@@ -2,10 +2,8 @@
 pub mod builtins;
 pub mod errors;
 pub mod local;
-pub mod proto;
 pub mod session;
 pub mod srv;
-pub mod types;
 pub mod validation;
 
 mod database;
