+++ crates/metastore/src/srv.rs
@@ -1,13 +1,13 @@
 use crate::database::DatabaseCatalog;
 use crate::errors::MetastoreError;
-use crate::proto::service::metastore_service_server::MetastoreService;
-use crate::proto::service::{
+use crate::storage::persist::Storage;
+use async_trait::async_trait;
+use metastoreproto::proto::service::metastore_service_server::MetastoreService;
+use metastoreproto::proto::service::{
     self, FetchCatalogRequest, FetchCatalogResponse, InitializeCatalogRequest,
     InitializeCatalogResponse, MutateRequest, MutateResponse,
 };
-use crate::storage::persist::Storage;
-use crate::types::service::Mutation;
-use async_trait::async_trait;
+use metastoreproto::types::service::Mutation;
 use object_store::ObjectStore;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -133,8 +133,8 @@ impl MetastoreService for Service {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::types::catalog::{CatalogEntry, CatalogState};
-    use crate::types::service::{CreateSchema, Mutation};
+    use metastoreproto::types::catalog::{CatalogEntry, CatalogState};
+    use metastoreproto::types::service::{CreateSchema, Mutation};
     use object_store::memory::InMemory;
 
     fn new_service() -> Service {
