+++ crates/metastore/src/storage/mod.rs
@@ -45,7 +45,7 @@ pub enum StorageError {
     LeaseRenewerExited,
 
     #[error(transparent)]
-    ProtoConv(#[from] crate::types::ProtoConvError),
+    ProtoConv(#[from] metastoreproto::types::ProtoConvError),
 
     #[error("Failed to encode protobuf for storage: {0}")]
     ProstEncode(#[from] prost::EncodeError),
