+++ crates/metastoreproto/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-name = "metastore"
+name = "metastoreproto"
 version = {workspace = true}
 edition = {workspace = true}
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-logutil = {path = "../logutil"}
-metastoreproto = { path = "../metastoreproto" }
-object_store_util = {path = "../object_store_util"}
-pgrepr = {path = "../pgrepr"}
 tonic = { version = "0.9", features = ["transport", "tls", "tls-roots"] }
 prost = "0.11"
 prost-types = "0.11"
