+++ crates/metastoreproto/proto/catalog.proto
