+++ crates/metastoreproto/src/lib.rs
@@ -1,0 +1,5 @@
+//! Protobuf types used in various crates.
+//!
+//! This exists outside the `metastore` to avoid dependency cycles.
+pub mod proto;
+pub mod types;
