+++ crates/metastoreproto/src/types/catalog.rs
