+++ crates/sqlexec/Cargo.toml
@@ -9,6 +9,7 @@ edition = {workspace = true}
 logutil = {path = "../logutil"}
 pgrepr = {path = "../pgrepr"}
 metastore = {path = "../metastore"}
+metastoreproto = { path = "../metastoreproto" }
 datafusion_planner = { path = "../datafusion_planner" }
 telemetry = {path = "../telemetry"}
 datasources = {path = "../datasources"}
