+++ crates/sqlexec/src/context.rs
@@ -22,8 +22,8 @@ use metastore::builtins::DEFAULT_CATALOG;
 use metastore::builtins::POSTGRES_SCHEMA;
 use metastore::errors::ResolveErrorStrategy;
 use metastore::session::SessionCatalog;
-use metastore::types::catalog::EntryType;
-use metastore::types::service::{self, Mutation};
+use metastoreproto::types::catalog::EntryType;
+use metastoreproto::types::service::{self, Mutation};
 use pgrepr::format::Format;
 use pgrepr::types::arrow_to_pg_type;
 use std::collections::HashMap;
